Modularization of daily item
in order to use it in multiple views
