+++ objetslitt/controllers/objetslitt_pages_controller.rb
@@ -1,7 +1,12 @@
+require_relative 'daily_item'
+
 class ObjetslittPagesController < PagesController
+    include DailyItem;
+
   def show
     @slug = request[:slug]
     @page = catalog.pages.where(:slug => @slug).first!
+    @daily_item_data = fetch_daily_item_data
 
     render :show
   end
