+++ objetslitt/views/catalogs/show.html.erb
@@ -48,36 +48,6 @@
     </div>
 </div>
 
-<div class="daily-item">
-    <% if @daily_item_cover %>
-        <div class="daily-item-cover">
-            <img src="<%= @daily_item_cover %>" class="daily-item-cover-image">
-        </div>
-    <% end %>
-
-    <div class="daily-item-content">
-        <h2 style="margin-top: 0;">Objet du jour</h2>
-        <h3 class="daily-item-title"><%= @daily_item_name %></h3>
-
-        <div style="width: 100%;">
-            <div class="daily-item-description">
-                <%= truncate(
-                    strip_tags(@daily_item_description),
-                    length: 400,
-                    separator: ' ',
-                    omission: '...'
-                ) %>
-            </div>
-
-            <div style="width: 75%;">
-                <a href="<%= @daily_item_url %>">
-                    <button class="daily-item-link btn btn-secondary">
-                        Voir la fiche
-                    </button>
-                </a>
-            </div>
-        </div>
-    </div>
-</div>
+<%= render partial: "shared/daily_item", locals: { daily_item_data: @daily_item_data } %>
 
 </div>
